+++ packages/angular-cli/blueprints/ng2/files/__path__/test.ts
@@ -1,3 +1,5 @@
+// This file is required by karma.conf.js and loads recursively all the .spec and framework files
+
 import './polyfills.ts';
 
 import 'zone.js/dist/long-stack-trace-zone';