+++ packages/angular-cli/commands/build.run.ts
@@ -1,6 +1,5 @@
 import { Version } from '../upgrade/version';
 import WebpackBuild from '../tasks/build-webpack';
-import WebpackBuildWatch from '../tasks/build-webpack-watch';
 import { BuildOptions } from './build';
 
 export default function buildRun(commandOptions: BuildOptions) {
@@ -22,23 +21,23 @@ export default function buildRun(commandOptions: BuildOptions) {
     }
   }
 
+  if (typeof commandOptions.sourcemap === 'undefined') {
+    if (commandOptions.target == 'development') {
+      commandOptions.sourcemap = true;
+    }
+    if (commandOptions.target == 'production') {
+      commandOptions.sourcemap = false;
+    }
+  }
+
   const project = this.project;
 
   // Check angular version.
   Version.assertAngularVersionIs2_3_1OrHigher(project.root);
 
-  const ui = this.ui;
-  const buildTask = commandOptions.watch ?
-    new WebpackBuildWatch({
+  const buildTask = new WebpackBuild({
       cliProject: project,
-      ui: ui,
-      outputPath: commandOptions.outputPath,
-      target: commandOptions.target,
-      environment: commandOptions.environment
-    }) :
-    new WebpackBuild({
-      cliProject: project,
-      ui: ui,
+      ui: this.ui,
       outputPath: commandOptions.outputPath,
       target: commandOptions.target,
       environment: commandOptions.environment,