+++ packages/angular-cli/models/webpack-config.ts
@@ -7,6 +7,7 @@ import { CliConfig } from './config';
 import { getWebpackCommonConfig } from './webpack-build-common';
 import { getWebpackDevConfigPartial } from './webpack-build-development';
 import { getWebpackProdConfigPartial } from './webpack-build-production';
+import { getWebpackStylesConfig } from './webpack-build-styles';
 import {
   getWebpackMobileConfigPartial,
   getWebpackMobileProdConfigPartial
@@ -33,16 +34,19 @@ export class NgCliWebpackConfig {
     verbose = false,
     progress = true,
     deployUrl?: string,
-    outputHashing?: string
+    outputHashing?: string,
+    extractCss = true,
   ) {
-    const config: CliConfig = CliConfig.fromProject();
-    const appConfig = config.config.apps[0];
+    const appConfig = CliConfig.fromProject().config.apps[0];
+    const projectRoot = this.ngCliProject.root;
 
+    appConfig.scripts = appConfig.scripts || [];
+    appConfig.styles = appConfig.styles || [];
     appConfig.outDir = outputDir || appConfig.outDir;
     appConfig.deployUrl = deployUrl || appConfig.deployUrl;
 
     let baseConfig = getWebpackCommonConfig(
-      this.ngCliProject.root,
+      projectRoot,
       environment,
       appConfig,
       baseHref,
@@ -50,30 +54,41 @@ export class NgCliWebpackConfig {
       vendorChunk,
       verbose,
       progress,
-      outputHashing
+      outputHashing,
     );
-    let targetConfigPartial = this.getTargetConfig(
-      this.ngCliProject.root, appConfig, sourcemap, verbose
-    );
-    const typescriptConfigPartial = isAoT
-      ? getWebpackAotConfigPartial(this.ngCliProject.root, appConfig, i18nFile, i18nFormat, locale)
-      : getWebpackNonAotConfigPartial(this.ngCliProject.root, appConfig);
+    let targetConfigPartial = this.getTargetConfig(projectRoot, appConfig, sourcemap, verbose);
 
     if (appConfig.mobile) {
-      let mobileConfigPartial = getWebpackMobileConfigPartial(this.ngCliProject.root, appConfig);
-      let mobileProdConfigPartial = getWebpackMobileProdConfigPartial(this.ngCliProject.root,
-                                                                      appConfig);
+      let mobileConfigPartial = getWebpackMobileConfigPartial(projectRoot, appConfig);
+      let mobileProdConfigPartial = getWebpackMobileProdConfigPartial(projectRoot, appConfig);
       baseConfig = webpackMerge(baseConfig, mobileConfigPartial);
       if (this.target == 'production') {
         targetConfigPartial = webpackMerge(targetConfigPartial, mobileProdConfigPartial);
       }
     }
 
-    this.config = webpackMerge(
-      baseConfig,
-      targetConfigPartial,
-      typescriptConfigPartial
+    let config = webpackMerge(baseConfig, targetConfigPartial);
+
+    if (appConfig.main) {
+      const typescriptConfigPartial = isAoT
+        ? getWebpackAotConfigPartial(projectRoot, appConfig, i18nFile, i18nFormat, locale)
+        : getWebpackNonAotConfigPartial(projectRoot, appConfig);
+
+      config = webpackMerge(config, typescriptConfigPartial);
+    }
+
+    const stylesConfig = getWebpackStylesConfig(
+      projectRoot,
+      appConfig,
+      target,
+      sourcemap,
+      outputHashing,
+      extractCss
     );
+
+    config = webpackMerge(config, stylesConfig);
+
+    this.config = config;
   }
 
   getTargetConfig(projectRoot: string, appConfig: any, sourcemap: boolean, verbose: boolean): any {