+++ tests/acceptance/github-pages-deploy.spec.js
@@ -51,6 +51,7 @@ describe('Acceptance: ng github-pages:deploy', function() {
   });
 
   it('should fail with uncommited changes', function() {
+    this.timeout(10000);
     execStub.addExecSuccess('git status --porcelain', 'M dir/file.ext');
     return ng(['github-pages:deploy', '--skip-build'])
       .then(() => {
@@ -75,6 +76,44 @@ describe('Acceptance: ng github-pages:deploy', function() {
       .addExecSuccess('git remote -v', remote);
 
     return ng(['github-pages:deploy', '--skip-build']);
+  });
+
+  it('should deploy with token and username', function () {
+    let token = 'token',
+      username = 'bar';
+
+    execStub.addExecSuccess('git status --porcelain')
+      .addExecSuccess('git rev-parse --abbrev-ref HEAD', initialBranch)
+      .addExecSuccess('git remote -v', remote)
+      .addExecSuccess(`git checkout ${ghPagesBranch}`)
+      .addExecSuccess('git ls-files')
+      .addExecSuccess('git rm -r ')
+      .addExecSuccess('git add .')
+      .addExecSuccess(`git commit -m "${message}"`)
+      .addExecSuccess(`git checkout ${initialBranch}`)
+      .addExecSuccess(`git push https://${token}@github.com/${username}/${project}.git ${ghPagesBranch}:${ghPagesBranch}`)
+      .addExecSuccess('git remote -v', remote);
+
+    return ng(['github-pages:deploy', '--skip-build', `--gh-token=${token}`, `--gh-username=${username}`]);
+  })
+
+  it('should deploy with token only', function () {
+    let token = 'token';
+
+    execStub.addExecSuccess('git status --porcelain')
+      .addExecSuccess('git rev-parse --abbrev-ref HEAD', initialBranch)
+      .addExecSuccess('git remote -v', remote)
+      .addExecSuccess(`git checkout ${ghPagesBranch}`)
+      .addExecSuccess('git ls-files')
+      .addExecSuccess('git rm -r ')
+      .addExecSuccess('git add .')
+      .addExecSuccess(`git commit -m "${message}"`)
+      .addExecSuccess(`git checkout ${initialBranch}`)
+      .addExecSuccess('git remote -v', remote)
+      .addExecSuccess(`git push https://${token}@github.com/username/${project}.git ${ghPagesBranch}:${ghPagesBranch}`)
+      .addExecSuccess('git remote -v', remote);
+
+    return ng(['github-pages:deploy', '--skip-build', `--gh-token=${token}`]);
   });
 
   it('should deploy with changed defaults', function() {
@@ -126,14 +165,14 @@ describe('Acceptance: ng github-pages:deploy', function() {
       .addExecSuccess('git rev-parse --abbrev-ref HEAD', initialBranch)
       .addExecSuccess('git remote -v', noRemote)
       .addExecSuccess(`git remote add origin git@github.com:${username}/${project}.git`)
-      .addExecSuccess(`git push -u origin ${initialBranch}`)
+      .addExecSuccess(`git push -u https://${token}@github.com/${username}/${project}.git ${initialBranch}`)
       .addExecSuccess(`git checkout ${ghPagesBranch}`)
       .addExecSuccess('git ls-files')
       .addExecSuccess('git rm -r ')
       .addExecSuccess('git add .')
       .addExecSuccess(`git commit -m "${message}"`)
       .addExecSuccess(`git checkout ${initialBranch}`)
-      .addExecSuccess(`git push origin ${ghPagesBranch}:${ghPagesBranch}`)
+      .addExecSuccess(`git push https://${token}@github.com/${username}/${project}.git ${ghPagesBranch}:${ghPagesBranch}`)
       .addExecSuccess('git remote -v', remote);
 
     var httpsStub = sinon.stub(https, 'request', httpsRequestStubFunc);