+++ CHANGELOG.md
@@ -1,3 +1,75 @@
+<a name="1.0.0-beta.26"></a>
+# [1.0.0-beta.26](https://github.com/angular/angular-cli/compare/v1.0.0-beta.25...v1.0.0-beta.26) (2017-01-19)
+
+
+### Bug Fixes
+
+* **@ngtools/json-schema:** serialize object properties better. ([#4103](https://github.com/angular/angular-cli/issues/4103)) ([48d1e44](https://github.com/angular/angular-cli/commit/48d1e44)), closes [#4044](https://github.com/angular/angular-cli/issues/4044)
+* **@ngtools/webpack:** dont error on non-identifier properties. ([#4078](https://github.com/angular/angular-cli/issues/4078)) ([e91552f](https://github.com/angular/angular-cli/commit/e91552f))
+* **@ngtools/webpack:** honor tsconfig#angularCompilerOptions.entryModule before trying to resolveEntryModuleFromMain() ([#4013](https://github.com/angular/angular-cli/issues/4013)) ([c9ac263](https://github.com/angular/angular-cli/commit/c9ac263))
+* **build:** override publicPath for ExtractTextPlugin and add extract-css test ([#4036](https://github.com/angular/angular-cli/issues/4036)) ([c1f1e0c](https://github.com/angular/angular-cli/commit/c1f1e0c))
+* **generate:** correct component path when module is generated in subfolder, and parent folder is not a module too ([#3916](https://github.com/angular/angular-cli/issues/3916)) ([f70feae](https://github.com/angular/angular-cli/commit/f70feae)), closes [#3255](https://github.com/angular/angular-cli/issues/3255)
+* **generate:** normalize pwd before using it ([#4065](https://github.com/angular/angular-cli/issues/4065)) ([09e1eb3](https://github.com/angular/angular-cli/commit/09e1eb3)), closes [#1639](https://github.com/angular/angular-cli/issues/1639)
+* **get/set:** Add support for global configuration. ([#4074](https://github.com/angular/angular-cli/issues/4074)) ([088ebf0](https://github.com/angular/angular-cli/commit/088ebf0))
+* **help:** remove ember references in console output ([#4026](https://github.com/angular/angular-cli/issues/4026)) ([394aa05](https://github.com/angular/angular-cli/commit/394aa05))
+* **help:** remove match of *.run.ts files ([#3982](https://github.com/angular/angular-cli/issues/3982)) ([7b47753](https://github.com/angular/angular-cli/commit/7b47753))
+* **new:** improve error message when project name does not match regex ([bf23b13](https://github.com/angular/angular-cli/commit/bf23b13)), closes [#3816](https://github.com/angular/angular-cli/issues/3816) [#3902](https://github.com/angular/angular-cli/issues/3902)
+* **test:** remove webpack size limit warning ([#3974](https://github.com/angular/angular-cli/issues/3974)) ([5df4799](https://github.com/angular/angular-cli/commit/5df4799))
+
+
+### Features
+
+* **@ngtools/json-schema:** add support for enums. ([c034a44](https://github.com/angular/angular-cli/commit/c034a44)), closes [#4082](https://github.com/angular/angular-cli/issues/4082)
+* **build:** add style paths ([#4003](https://github.com/angular/angular-cli/issues/4003)) ([e5ef996](https://github.com/angular/angular-cli/commit/e5ef996)), closes [#1791](https://github.com/angular/angular-cli/issues/1791)
+* **build:** use NamedModulesPlugin with HMR ([c5b2244](https://github.com/angular/angular-cli/commit/c5b2244)), closes [#3679](https://github.com/angular/angular-cli/issues/3679) [#4037](https://github.com/angular/angular-cli/issues/4037)
+* **deploy:** add custom-domain support for gh-pages deployment ([#1781](https://github.com/angular/angular-cli/issues/1781)) ([#3392](https://github.com/angular/angular-cli/issues/3392)) ([a54bc16](https://github.com/angular/angular-cli/commit/a54bc16))
+* **generate:** add option to auto-export declarations ([#3876](https://github.com/angular/angular-cli/issues/3876)) ([6d63bb4](https://github.com/angular/angular-cli/commit/6d63bb4)), closes [#3778](https://github.com/angular/angular-cli/issues/3778)
+* **generate:** create parent directories required for blueprints if they do not exist ([76380a6](https://github.com/angular/angular-cli/commit/76380a6)), closes [#3307](https://github.com/angular/angular-cli/issues/3307) [#3912](https://github.com/angular/angular-cli/issues/3912)
+* **gh-pages:deploy:** add aot and vendor-chunk options for gh-pages:deploy ([#4073](https://github.com/angular/angular-cli/issues/4073)) ([71445c3](https://github.com/angular/angular-cli/commit/71445c3))
+
+
+
+<a name="1.0.0-beta.25"></a>
+# [1.0.0-beta.25](https://github.com/angular/angular-cli/compare/v1.0.0-beta.24...v1.0.0-beta.25) (2017-01-12)
+
+
+### Bug Fixes
+
+* **@ngtools/webpack:** fix tsconfig paths resolver ([#3831](https://github.com/angular/angular-cli/issues/3831)) ([c6d1c99](https://github.com/angular/angular-cli/commit/c6d1c99)), closes [#3586](https://github.com/angular/angular-cli/issues/3586)
+* **@ngtools/webpack:** search recursively for entry module ([#3708](https://github.com/angular/angular-cli/issues/3708)) ([bb748f2](https://github.com/angular/angular-cli/commit/bb748f2))
+* **build:** close tags in index.html ([#3743](https://github.com/angular/angular-cli/issues/3743)) ([aaca100](https://github.com/angular/angular-cli/commit/aaca100)), closes [#3217](https://github.com/angular/angular-cli/issues/3217)
+* **build:** disable performance hints ([#3808](https://github.com/angular/angular-cli/issues/3808)) ([2a513ca](https://github.com/angular/angular-cli/commit/2a513ca))
+* **build:** fix path error when appConfig has no main ([#3867](https://github.com/angular/angular-cli/issues/3867)) ([7bd165b](https://github.com/angular/angular-cli/commit/7bd165b))
+* **build/serve:** correct check against angular v2.3.1 ([#3785](https://github.com/angular/angular-cli/issues/3785)) ([d0224a5](https://github.com/angular/angular-cli/commit/d0224a5)), closes [#3720](https://github.com/angular/angular-cli/issues/3720) [#3729](https://github.com/angular/angular-cli/issues/3729)
+* **config:** allow minimal config for build/serve ([#3835](https://github.com/angular/angular-cli/issues/3835)) ([f616158](https://github.com/angular/angular-cli/commit/f616158))
+* **lint:** remove tslint rule that requires type info ([#3818](https://github.com/angular/angular-cli/issues/3818)) ([1555c2b](https://github.com/angular/angular-cli/commit/1555c2b))
+* **lint:** use noUnusedParameters and noUnusedLocals instead of no-unused-variable ([#3945](https://github.com/angular/angular-cli/issues/3945)) ([dd378fe](https://github.com/angular/angular-cli/commit/dd378fe))
+* **serve:** communicate that ng serve is not secure. ([#3646](https://github.com/angular/angular-cli/issues/3646)) ([766394d](https://github.com/angular/angular-cli/commit/766394d))
+* **serve:** fallback to config.app[0].index ([#3813](https://github.com/angular/angular-cli/issues/3813)) ([45e2985](https://github.com/angular/angular-cli/commit/45e2985)), closes [#3748](https://github.com/angular/angular-cli/issues/3748)
+* **test:** remove webpack size limit warning ([#3974](https://github.com/angular/angular-cli/issues/3974)) ([b25b97d](https://github.com/angular/angular-cli/commit/b25b97d))
+* **tests:** add global scripts in karma plugin ([#3543](https://github.com/angular/angular-cli/issues/3543)) ([1153c92](https://github.com/angular/angular-cli/commit/1153c92)), closes [#2897](https://github.com/angular/angular-cli/issues/2897)
+
+
+### Features
+
+* **@ngtools/json-schema:** Introduce a separate package for JSON schema. ([#3927](https://github.com/angular/angular-cli/issues/3927)) ([74f7cdd](https://github.com/angular/angular-cli/commit/74f7cdd))
+* **@ngtools/logger:** Implement a reactive logger. ([#3774](https://github.com/angular/angular-cli/issues/3774)) ([e3b48da](https://github.com/angular/angular-cli/commit/e3b48da))
+* **@ngtools/webpack:** convert dashless resource urls ([#3842](https://github.com/angular/angular-cli/issues/3842)) ([4e7b397](https://github.com/angular/angular-cli/commit/4e7b397))
+* **build:** add --extract-css flag ([#3943](https://github.com/angular/angular-cli/issues/3943)) ([87536c8](https://github.com/angular/angular-cli/commit/87536c8))
+* **build:** add publicPath support via command and angular-cli.json ([#3285](https://github.com/angular/angular-cli/issues/3285)) ([0ce64a4](https://github.com/angular/angular-cli/commit/0ce64a4)), closes [#3136](https://github.com/angular/angular-cli/issues/3136) [#2960](https://github.com/angular/angular-cli/issues/2960) [#2276](https://github.com/angular/angular-cli/issues/2276) [#2241](https://github.com/angular/angular-cli/issues/2241) [#3344](https://github.com/angular/angular-cli/issues/3344)
+* **build:** allow output hashing to be configured ([#3885](https://github.com/angular/angular-cli/issues/3885)) ([b82fe41](https://github.com/angular/angular-cli/commit/b82fe41))
+* **build:** disable sourcemaps for production ([#3963](https://github.com/angular/angular-cli/issues/3963)) ([da1c197](https://github.com/angular/angular-cli/commit/da1c197))
+* **commands:** lazy load commands ([#3805](https://github.com/angular/angular-cli/issues/3805)) ([59e9e8f](https://github.com/angular/angular-cli/commit/59e9e8f))
+* **deploy:github-pages:** support usage of gh-token for deployment from external env ([#3121](https://github.com/angular/angular-cli/issues/3121)) ([3c82b77](https://github.com/angular/angular-cli/commit/3c82b77))
+* **generate:** add ability to specify module for import ([#3811](https://github.com/angular/angular-cli/issues/3811)) ([e2b051f](https://github.com/angular/angular-cli/commit/e2b051f)), closes [#3806](https://github.com/angular/angular-cli/issues/3806)
+* **lint:** now lint e2e ts files as well ([#3941](https://github.com/angular/angular-cli/issues/3941)) ([f84e220](https://github.com/angular/angular-cli/commit/f84e220))
+* **new:** add --skip-tests flag to ng new/init to skip creating spec files ([#3825](https://github.com/angular/angular-cli/issues/3825)) ([4c2f06a](https://github.com/angular/angular-cli/commit/4c2f06a))
+* **new:** add flag to prevent initial git commit ([#3712](https://github.com/angular/angular-cli/issues/3712)) ([2e2377d](https://github.com/angular/angular-cli/commit/2e2377d))
+* **new:** show name of created project in output ([#3795](https://github.com/angular/angular-cli/issues/3795)) ([888beb7](https://github.com/angular/angular-cli/commit/888beb7))
+* **version:** compare local and global version and warn users. ([#3693](https://github.com/angular/angular-cli/issues/3693)) ([8b47a90](https://github.com/angular/angular-cli/commit/8b47a90))
+
+
+
 <a name="1.0.0-beta.24"></a>
 # [1.0.0-beta.24](https://github.com/angular/angular-cli/compare/v1.0.0-beta.23...v1.0.0-beta.24) (2016-12-20)